🧰🧑‍🏫: reduce build size of classes runtime build
