+++ lively.ide/js/objecteditor/classTree.js
@@ -1,7 +1,7 @@
 import { TreeData } from 'lively.components';
 import { arr } from 'lively.lang';
 import { withSuperclasses } from 'lively.classes/util.js';
-import { lexicalClassMembers } from 'lively.classes/util.js';
+import { lexicalClassMembers } from 'lively.classes/source-descriptors.js';
 import { Icon } from 'lively.morphic';
 
 // var oe = ObjectEditor.open({target: this})
